+++ exercicios/src/CalculadoraMultiplicacao.java
@@ -1,6 +1,6 @@
 public class CalculadoraMultiplicacao {
-    public int main(int num1,int num2){
-        int multiplicacao = num1 * num2;
+    public double main(double num1,double num2){
+        double multiplicacao = num1 * num2;
         return multiplicacao;
     }
 }
