+++ exercicios/src/CalculadoraSoma.java
@@ -1,6 +1,6 @@
 public class CalculadoraSoma {
-    public int main(int num1, int num2) {
-        int soma = num1 + num2;
+    public double main(double num1, double num2) {
+        double soma = num1 + num2;
         return soma;
     }
 }
